requirements/dev.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
charset-normalizer==3.1.0
dnspython==2.3.0
email-validator==2.0.0.post2
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pydantic==1.10.8
PyJWT==2.7.0
typing_extensions==4.6.2
yarl==1.9.2
